requirements.txt
dill==0.3.5.1
dnspython==2.2.1
fastapi==0.78.0
httptools==0.5.0
imblearn==0.0
scikit-learn==1.3.2
mypy-boto3-s3==1.24.76
pip-chill==1.0.1
pymongo==4.2.0
setuptools==75.1.0
typing==3.7.4.3
python-dotenv==1.0.1
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==2.1.4
pandas==2.0.3
PyYAML==6.0.2
numpy==1.24.4
scipy==1.10.1
apache-airflow==2.10.5
psycopg2-binary>=2.9,<3.0
apache-airflow-providers-amazon==8.29.0
-e .